requirements.txt
streamlit
groq
google-generativeai
matplotlib